Simplify API by removing an alternate SM constructor
In my opinion, the reflected constructor was a middle road solution -
It required users understand that StateReflector exists, and what it
does (roughly), but didn't require them to know the *only* method in
that class, GetStates(). It should either new up a reflector, and be
dependent on it, or be entirely independent of the IStateReflector
class. I chose to do the latter since it's easier to test and maintain,
and only asks users to know a bit more about the reflection utility.
It's also more explicit - calling code with be more declarative, because
it won't hide the reflection call inside the constructor.